+++ src/main/java/org/takintosh/domain/models/Appointment.java
@@ -3,12 +3,12 @@ package org.takintosh.domain.models;
 import java.time.LocalDateTime;
 
 public class Appointment {
-    private Long id;
+    private int id;
     private LocalDateTime date;
     private Client client;
     private Barber barber;
 
-    public Appointment(Long id, LocalDateTime date, Client client, Barber barber) {
+    public Appointment(int id, LocalDateTime date, Client client, Barber barber) {
         this.id = id;
         this.date = date;
         this.client = client;
@@ -16,10 +16,10 @@ public class Appointment {
     }
 
     // Getters and setters
-    public Long getId() {
+    public int getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
